==> 3 polygon.py <==
import folium
import webbrowser

#описание полигона Билимбая
polygon_bilimnay = '{"type": "FeatureCollection", "features": ' \
                   '[{"id": "974", "type": "Feature", "properties": {}, "geometry": {"type": "Polygon", "coordinates": ' \
                   '[[[59.8052902, 56.983266], [59.8169303, 56.9890487], [59.8237753, 56.989037], [59.8258449, 56.9813634], ' \
                   '[59.822912, 56.9765333], [59.8385601, 56.9732184], [59.8528097, 56.9641286], [59.8525307, 56.9615015], ' \
                   '[59.8484428, 56.9602115], [59.8360969, 56.9498693], [59.8302966, 56.9493654], [59.8280542, 56.934437], ' \
                   '[59.8064534, 56.9364719], [59.797701, 56.9358102], [59.7788918, 56.9436259], [59.7911249, 56.9479611], ' \
                   '[59.7973257, 56.9463994], [59.8020703, 56.9504127], [59.7931892, 56.9573773], [59.7957583, 56.9723638], ' \
                   '[59.8052902, 56.983266]]]}, "bbox": [59.7788918, 56.934437, 59.8528097, 56.9890487]}], "bbox": ' \
                   '[59.7788918, 56.934437, 59.8528097, 56.9890487]}'

m = folium.Map(location=[56.962230, 59.818010], zoom_start=12, tiles='OpenStreetMap')

# создаем полигон
geo_bilimbay = folium.GeoJson(data=polygon_bilimnay,
                              tooltip= 'Свердловская область<br>городской округ Первоуральск<br>п. Билимбай', #подсказки всплывающие при наведении мыши
                              style_function=lambda x: {'fillColor': 'red', 'fillOpacity': 0.1, 'color': 'blue', 'weight': 3, 'opacity': 0.9})

# добавляем статическую подсказку по клику
folium.Popup('Численность населения — 5 875 чел.<br>').add_to(geo_bilimbay) # подсказки статика по клику мыши

# размещаем полигон на корте
geo_bilimbay.add_to(m)

#отображаем карту
m.save('results/map.html')
webbrowser.open('/results/map.html')
