==> 4 polygon.py <==
import folium
import webbrowser

#описание полигона Еката
polygon_ekb = '{"type": "FeatureCollection", "features": ' \
              '[{"id": "1532", "type": "Feature", "properties": {}, "geometry": {"type": "Polygon", "coordinates": ' \
              '[[[60.1921821, 56.7829028], [60.3106795, 56.771934], [60.330996, 56.8352299], [60.3260406, 56.8353889], ' \
              '[60.3226653, 56.8397394], [60.3144136, 56.8443147], [60.2906042, 56.8447415], [60.2710873, 56.8478982], ' \
              '[60.265983, 56.8544176], [60.2605434, 56.853676], [60.2694515, 56.8812545], [60.2716448, 56.8811598], ' \
              '[60.2794919, 56.9099876], [60.3118899, 56.9072965], [60.3265171, 56.9647026], [60.3608655, 56.96158], ' \
              '[60.3557562, 56.9425235], [60.373537, 56.9406161], [60.3718486, 56.9319104], [60.403121, 56.9269218], ' \
              '[60.4055719, 56.9227348], [60.4081412, 56.9226739], [60.4135684, 56.9190294], [60.4672172, 56.9155446], ' \
              '[60.4705458, 56.9191095], [60.4719781, 56.9179911], [60.4699584, 56.9172436], [60.4712174, 56.9159919], ' \
              '[60.4781827, 56.9147471], [60.4854298, 56.9187434], [60.4757672, 56.922515], [60.4742761, 56.9234632], ' \
              '[60.476182, 56.9238014], [60.4729172, 56.925768], [60.485025, 56.9306884], [60.5045955, 56.9346919], ' \
              '[60.5111219, 56.9394534], [60.5162229, 56.9371318], [60.5308748, 56.9401122], [60.5776317, 56.9432883], ' \
              '[60.6105919, 56.9410529], [60.6135374, 56.9415905], [60.613762, 56.9430421], [60.6250342, 56.9441552], ' \
              '[60.6567448, 56.9434004], [60.6716743, 56.9588279], [60.6698063, 56.9619959], [60.6706782, 56.9684037], ' \
              '[60.6694378, 56.9757789], [60.6885982, 56.9740721], [60.6913204, 56.9826914], [60.7249895, 56.9792213], ' \
              '[60.7233779, 56.9743294], [60.7364528, 56.9742369], [60.7340607, 56.9701316], [60.7402206, 56.9676051], ' \
              '[60.7417754, 56.9704963], [60.7478505, 56.9699948], [60.750093, 56.9757957], [60.7438391, 56.9753451], ' \
              '[60.7492195, 56.9766165], [60.7586146, 56.9756078], [60.7548064, 56.9651714], [60.7577086, 56.9648602], ' \
              '[60.7567681, 56.9622445], [60.7541596, 56.9605185], [60.7530518, 56.9609101], [60.7536111, 56.9637868], ' \
              '[60.7520496, 56.9634315], [60.7533287, 56.9629945], [60.7517843, 56.9607209], [60.754574, 56.9581825], ' \
              '[60.753278, 56.9570085], [60.7366278, 56.9580952], [60.7315661, 56.9549397], [60.7197746, 56.9518945], ' \
              '[60.7163593, 56.9486996], [60.7225738, 56.9347093], [60.7386808, 56.9303639], [60.7392762, 56.9286469], ' \
              '[60.7380857, 56.9223187], [60.739482, 56.9211626], [60.7363608, 56.9193236], [60.7400266, 56.9188681], ' \
              '[60.7355156, 56.9055457], [60.7313009, 56.9051783], [60.7352007, 56.9047793], [60.7348337, 56.9034583], ' \
              '[60.7235787, 56.8999096], [60.7232046, 56.892009], [60.7463499, 56.8894781], [60.7418287, 56.8730339], ' \
              '[60.7532892, 56.8732746], [60.7732932, 56.8690651], [60.7733279, 56.8673726], [60.7811537, 56.868584], ' \
              '[60.7874541, 56.8654899], [60.7918383, 56.8659593], [60.7896139, 56.8622119], [60.771567, 56.8601397], ' \
              '[60.770968, 56.857872], [60.9432814, 56.8402997], [60.9148065, 56.7646111], [60.9185512, 56.7667985], ' \
              '[60.9209357, 56.7652012], [60.9187326, 56.7645519], [60.9197967, 56.7635356], [60.9118718, 56.7617209], ' \
              '[60.90879, 56.7633393], [60.8971903, 56.7615175], [60.8893254, 56.7547903], [60.8835549, 56.7534303], ' \
              '[60.8789814, 56.738391], [60.8764465, 56.7379272], [60.8655657, 56.7258177], [60.8658864, 56.7245813], ' \
              '[60.8714309, 56.7238773], [60.8901478, 56.7252255], [60.9043905, 56.718559], [60.914018, 56.7164397], ' \
              '[60.9027762, 56.7123556], [60.8735583, 56.6957431], [60.8655121, 56.7028443], [60.865254, 56.705171], ' \
              '[60.86251, 56.7054351], [60.8449266, 56.7211237], [60.8297748, 56.7197272], [60.8207753, 56.7158991], ' \
              '[60.7907661, 56.7288501], [60.7851677, 56.7299081], [60.7853282, 56.7320998], [60.779178, 56.7338694], ' \
              '[60.7783012, 56.739225], [60.7724464, 56.7414598], [60.7733448, 56.7452504], [60.7686159, 56.7464685], ' \
              '[60.760025, 56.7412943], [60.7508458, 56.7323485], [60.7429448, 56.7329127], [60.7415738, 56.7355905], ' \
              '[60.737279, 56.7361557], [60.7335638, 56.7386987], [60.7194319, 56.7306871], [60.7240632, 56.7284752], ' \
              '[60.7370832, 56.7274001], [60.7374621, 56.7226206], [60.7428066, 56.7165525], [60.7446355, 56.7163516], ' \
              '[60.7379118, 56.7120167], [60.7484985, 56.7034642], [60.7520476, 56.6958603], [60.7650372, 56.6947964], ' \
              '[60.7650591, 56.6969132], [60.7673606, 56.6968935], [60.7674587, 56.7002454], [60.7687899, 56.7007835], ' \
              '[60.783967, 56.7010393], [60.7839134, 56.699809], [60.779072, 56.6996368], [60.7788297, 56.6947215], ' \
              '[60.7604229, 56.6945545], [60.7503646, 56.6955156], [60.7363803, 56.7000798], [60.7366684, 56.697774], ' \
              '[60.6569472, 56.6744084], [60.6553953, 56.6512843], [60.6110095, 56.6533043], [60.6080016, 56.6471233], ' \
              '[60.6000501, 56.6490744], [60.5944642, 56.654344], [60.5983716, 56.6439818], [60.5935591, 56.6407903], ' \
              '[60.5897719, 56.6428265], [60.5903061, 56.6457737], [60.5801075, 56.6473539], [60.577484, 56.6440945], ' \
              '[60.5701757, 56.6456529], [60.5716918, 56.6476733], [60.5673792, 56.6495949], [60.5601099, 56.6436693], ' \
              '[60.5734305, 56.6405254], [60.5678576, 56.6358819], [60.5846289, 56.6274193], [60.5815742, 56.6234305], ' \
              '[60.5920848, 56.6270908], [60.5903466, 56.6317973], [60.5917844, 56.633201], [60.6053972, 56.6345753], ' \
              '[60.6000049, 56.6386558], [60.6013091, 56.6404901], [60.5991349, 56.6422682], [60.6050136, 56.6427908], ' \
              '[60.6077718, 56.6467033], [60.6229168, 56.6440907], [60.6204898, 56.6370737], [60.6162355, 56.6348399], ' \
              '[60.6197178, 56.6316601], [60.6149822, 56.6341455], [60.6092486, 56.630708], [60.6141667, 56.6279696], ' \
              '[60.6204429, 56.6311222], [60.6297982, 56.621945], [60.6524955, 56.6193784], [60.64301, 56.6065726], ' \
              '[60.6020905, 56.6167885], [60.593672, 56.621772], [60.5893571, 56.6201172], [60.5815001, 56.62332], ' \
              '[60.5765943, 56.609578], [60.5841152, 56.6078704], [60.5798175, 56.6056291], [60.5730641, 56.6065877], ' \
              '[60.5617661, 56.6031838], [60.5559003, 56.604966], [60.5587717, 56.6136954], [60.5432867, 56.6153818], [60.5419431, 56.6119679], [60.5347802, 56.6127368], [60.5329923, 56.6078698], [60.5434672, 56.598852], [60.5429766, 56.5948808], [60.5413486, 56.5937449], [60.5374381, 56.5997017], [60.5312609, 56.6030503], [60.5083751, 56.6054286], [60.523061, 56.6070114], [60.519906, 56.6111732], [60.5000048, 56.6082895], [60.4976683, 56.6031079], [60.4382886, 56.6156686], [60.4488638, 56.6261335], [60.408012, 56.6433551], [60.4125558, 56.6473839], [60.4085045, 56.6487808], [60.4340232, 56.6625345], [60.4289164, 56.6641142], [60.429869, 56.6683677], [60.4333552, 56.671293], [60.4203217, 56.6701552], [60.407181, 56.6649749], [60.3447189, 56.6538037], [60.3385442, 56.6542085], [60.3332674, 56.6567592], [60.3221831, 56.6548792], [60.3202675, 56.6567147], [60.3218757, 56.6581514], [60.3226308, 56.6630959], [60.3147578, 56.6660959], [60.2935483, 56.6693107], [60.2896607, 56.6731554], [60.2912997, 56.6735726], [60.2899636, 56.6785469], [60.2848445, 56.6831759], [60.2687166, 56.6828212], [60.2644253, 56.6850898], [60.2639088, 56.6868329], [60.2672209, 56.689979], [60.2658626, 56.690404], [60.2709722, 56.6946106], [60.2688297, 56.6969724], [60.2623162, 56.6963055], [60.2628481, 56.7015707], [60.2533938, 56.7101617], [60.2554586, 56.7149181], [60.2354206, 56.7219463], [60.2047032, 56.7263462], [60.1874698, 56.7329771], [60.1814396, 56.7309484], [60.1663538, 56.7423799], [60.1690114, 56.7424524], [60.1693919, 56.7440105], [60.1669243, 56.7499125], [60.1631022, 56.7499833], [60.1557815, 56.7533805], [60.152349, 56.7567668], [60.140716, 56.7607537], [60.1390166, 56.7630539], [60.1351156, 56.7632294], [60.1182969, 56.7708192], [60.1167898, 56.7731666], [60.1193423, 56.7748626], [60.115685, 56.77652], [60.1138921, 56.7797516], [60.102567, 56.7782918], [60.1006266, 56.7800066], [60.0963374, 56.7782332], [60.0878782, 56.780821], [60.0831478, 56.7860875], [60.0765477, 56.7895832], [60.0715015, 56.7894904], [60.0696961, 56.7873795], [60.0701356, 56.7842481], [60.0674602, 56.7833534], [60.0595365, 56.7830745], [60.0571826, 56.7842959], [60.0564215, 56.7903597], [60.0541219, 56.7909191], [60.0588075, 56.7953151], [60.064511, 56.7962058], [60.0622544, 56.7995747], [60.0635377, 56.8026291], [60.0731764, 56.8001966], [60.0778756, 56.8033469], [60.0706385, 56.8071836], [60.0721032, 56.8090573], [60.0708041, 56.8109268], [60.0612207, 56.8143014], [60.0602489, 56.8104321], [60.063274, 56.8072796], [60.0612949, 56.8060214], [60.0562543, 56.8068447], [60.0556876, 56.80505], [60.0446176, 56.809511], [60.0391288, 56.8060384], [60.0339344, 56.8050207], [60.0250767, 56.8063477], [60.0208085, 56.8032924], [60.0237834, 56.7988181], [60.0115651, 56.8023556], [60.0070815, 56.8058752], [60.0109298, 56.8077801], [60.0142769, 56.8048011], [60.0214781, 56.8099031], [60.0423405, 56.8127042], [60.0501788, 56.8128592], [60.0538755, 56.8110077], [60.0604845, 56.8205912], [60.0705443, 56.8237558], [60.086391, 56.8195581], [60.0905637, 56.8202001], [60.0898527, 56.819261], [60.0913777, 56.8190052], [60.0891902, 56.8180019], [60.0898372, 56.8169929], [60.0943873, 56.8172746], [60.0998379, 56.8209434], [60.1031136, 56.8208278], [60.1091301, 56.8188159], [60.1052248, 56.8145489], [60.1051002, 56.8119703], [60.1078766, 56.8109189], [60.1112718, 56.8114076], [60.1153167, 56.8152388], [60.1203107, 56.8155531], [60.124536, 56.8129917], [60.1347498, 56.8136494], [60.1280875, 56.8113527], [60.1284072, 56.811304], [60.1384871, 56.814074], [60.1290799, 56.8112852], [60.1287548, 56.8108949], [60.1385754, 56.813563], [60.1317514, 56.8114793], [60.1302606, 56.8103097], [60.1272651, 56.8105602], [60.1284125, 56.8101024], [60.1299376, 56.8100827], [60.1375315, 56.8129827], [60.143624, 56.8123411], [60.1462199, 56.8095233], [60.1445575, 56.8079054], [60.1471395, 56.8056478], [60.1536569, 56.8047616], [60.1534805, 56.8032751], [60.1492887, 56.8017613], [60.1542687, 56.8014705], [60.1518959, 56.800191], [60.1564505, 56.7979623], [60.157729, 56.7940509], [60.1542079, 56.7918404], [60.1598201, 56.7884989], [60.1578536, 56.7835833], [60.1603702, 56.7818088], [60.1593735, 56.7807714], [60.165503, 56.7805121], [60.1656842, 56.7779273], [60.1638585, 56.7760071], [60.1678352, 56.7733888], [60.172573, 56.7746478], [60.1737119, 56.774056], [60.172668, 56.772254], [60.1798317, 56.7721505], [60.1846232, 56.7758675], [60.1857604, 56.7795869], [60.1921821, 56.7829028]], [[60.4907718, 56.9180307], [60.4926473, 56.9235792], [60.4834467, 56.9258337], [60.4807229, 56.9256382], [60.479, 56.9232476], [60.4760902, 56.9234183], [60.4907718, 56.9180307]], [[60.6774581, 56.9426732], [60.6786206, 56.9438395], [60.6851876, 56.9427983], [60.6868221, 56.9441224], [60.6855151, 56.9460099], [60.67772, 56.9482337], [60.6804165, 56.9510327], [60.6838082, 56.9501373], [60.6847626, 56.9514241], [60.6824172, 56.9532868], [60.6881067, 56.9604485], [60.68402, 56.9631799], [60.6714346, 56.9648339], [60.6702144, 56.9621057], [60.6718133, 56.9594963], [60.6718404, 56.9584486], [60.6635288, 56.9497949], [60.6745822, 56.9490015], [60.6772037, 56.9460313], [60.6746587, 56.9432152], [60.6774581, 56.9426732]], [[60.2876097, 56.8596324], [60.2978717, 56.8623838], [60.3084442, 56.8622032], [60.3100471, 56.8634669], [60.3021601, 56.8642603], [60.3035272, 56.8684189], [60.299742, 56.8706779], [60.3053025, 56.8719684], [60.3038464, 56.8738693], [60.3045828, 56.8774563], [60.2915704, 56.8787116], [60.2909097, 56.8775878], [60.2965511, 56.8760684], [60.2895433, 56.8757557], [60.2836543, 56.8771911], [60.2822569, 56.8742935], [60.2860401, 56.8738494], [60.2846438, 56.8709555], [60.2902771, 56.8699748], [60.2895054, 56.8682863], [60.2825045, 56.8678042], [60.2809021, 56.8614262], [60.284687, 56.8612217], [60.2851627, 56.8589008], [60.2876097, 56.8596324]], [[60.4973373, 56.714759], [60.5010281, 56.7164968], [60.4951784, 56.7183209], [60.4907567, 56.7158049], [60.4973373, 56.714759]], [[60.5845993, 56.6738706], [60.5886739, 56.6747157], [60.589763, 56.6806001], [60.5806669, 56.6815158], [60.5775034, 56.6762088], [60.5845993, 56.6738706]], [[60.3838242, 56.8456424], [60.3883494, 56.8494378], [60.3855529, 56.8498574], [60.3874925, 56.8528929], [60.3860277, 56.8550014], [60.3747232, 56.856782], [60.377107, 56.8486567], [60.3800316, 56.8494563], [60.3791885, 56.8463387], [60.3838242, 56.8456424]], [[60.4208847, 56.7781234], [60.4243363, 56.7818028], [60.417118, 56.7852029], [60.4068765, 56.7855119], [60.4036468, 56.782961], [60.4208847, 56.7781234]], [[60.5115697, 56.7142706], [60.5152001, 56.7144318], [60.515924, 56.7197861], [60.5125803, 56.7203506], [60.5139706, 56.7169276], [60.5102244, 56.7153532], [60.5115697, 56.7142706]], [[60.292305, 56.8546928], [60.2979271, 56.856475], [60.295137, 56.8585464], [60.300247, 56.8585321], [60.3017948, 56.8603008], [60.3154473, 56.8605239], [60.2974695, 56.861606], [60.2802445, 56.8561302], [60.2841032, 56.8552216], [60.2909345, 56.8564501], [60.292305, 56.8546928]], [[60.2772747, 56.8515705], [60.2816519, 56.8525673], [60.2811375, 56.8547605], [60.2760692, 56.853364], [60.2772747, 56.8515705]], [[60.5999345, 56.6246078], [60.600675, 56.6262236], [60.5967276, 56.6265333], [60.5967211, 56.6248626], [60.5999345, 56.6246078]], [[60.8903433, 56.7703555], [60.8964642, 56.7773534], [60.9059461, 56.7754024], [60.9129321, 56.7806229], [60.9145021, 56.7849981], [60.9092066, 56.7855216], [60.9100014, 56.7905263], [60.9089179, 56.7909469], [60.8571939, 56.7960098], [60.8552692, 56.790657], [60.8578805, 56.7879413], [60.8568322, 56.7871817], [60.8610591, 56.7844291], [60.865627, 56.7852291], [60.8581144, 56.7809786], [60.8608422, 56.7762774], [60.86873, 56.7749519], [60.8751529, 56.7760634], [60.8757667, 56.7733262], [60.8903433, 56.7703555]], [[60.3780098, 56.7673516], [60.3807784, 56.7675229], [60.3826136, 56.771452], [60.3852983, 56.7720589], [60.3872736, 56.7750727], [60.3711408, 56.779082], [60.3733051, 56.7761531], [60.3643828, 56.7743872], [60.3687056, 56.7710813], [60.3687444, 56.7694723], [60.3713495, 56.7702368], [60.3740425, 56.7679127], [60.3765424, 56.768586], [60.3780098, 56.7673516]], [[60.8776711, 56.7630211], [60.8732164, 56.7686804], [60.8747972, 56.7685251], [60.8751344, 56.7711052], [60.8832859, 56.7713347], [60.8728753, 56.7733451], [60.8685063, 56.7677571], [60.8696302, 56.7655784], [60.8672511, 56.7641747], [60.8776711, 56.7630211]], [[60.5684886, 56.6643629], [60.5732185, 56.6726461], [60.5777945, 56.6787772], [60.5726313, 56.6786118], [60.5714507, 56.6759873], [60.5729581, 56.6757041], [60.5710099, 56.6732423], [60.5731179, 56.6726706], [60.5684886, 56.6643629]], [[60.2954006, 56.7558103], [60.3027769, 56.7574478], [60.3052511, 56.7634449], [60.3078261, 56.7640886], [60.3082367, 56.7660034], [60.2959491, 56.769687], [60.2926256, 56.7691448], [60.2913554, 56.7653207], [60.2832791, 56.7648501], [60.288739, 56.7616275], [60.2879229, 56.7550943], [60.2937177, 56.7546176], [60.2954006, 56.7558103]], [[60.5364543, 56.7520097], [60.5404908, 56.7544595], [60.5410877, 56.7564322], [60.5484407, 56.7557002], [60.5491977, 56.7582452], [60.5574014, 56.7570539], [60.5559284, 56.7535986], [60.5683372, 56.7525065], [60.5704939, 56.7582037], [60.5683688, 56.7584166], [60.5690088, 56.7601129], [60.5759271, 56.7598504], [60.5799213, 56.761621], [60.5700588, 56.7643778], [60.5651915, 56.7680074], [60.5586011, 56.7653541], [60.5524591, 56.7665179], [60.5440536, 56.7639539], [60.5438592, 56.7670452], [60.538742, 56.7691504], [60.5370491, 56.7677907], [60.5409932, 56.762617], [60.5314626, 56.7598942], [60.5341852, 56.7560397], [60.5332709, 56.7529775], [60.5364543, 56.7520097]], [[60.4696448, 56.6632272], [60.4829462, 56.6672443], [60.4796671, 56.6681792], [60.4801165, 56.6691116], [60.4843346, 56.6693454], [60.4868381, 56.6680124], [60.488297, 56.6678388], [60.4889542, 56.6679567], [60.4844128, 56.6694771], [60.4795251, 56.673867], [60.4620897, 56.6668578], [60.4651228, 56.6634428], [60.4696448, 56.6632272]], [[60.2739058, 56.7088516], [60.2868115, 56.7099375], [60.286913, 56.7143109], [60.2808753, 56.7148127], [60.2807789, 56.7175218], [60.277483, 56.7178802], [60.2772052, 56.7152368], [60.2672651, 56.7144091], [60.2691001, 56.7123725], [60.266505, 56.7104662], [60.2735601, 56.7111422], [60.2739058, 56.7088516]], [[60.5088486, 56.8002759], [60.5107995, 56.8028207], [60.5036685, 56.8046259], [60.5061071, 56.8076909], [60.4976835, 56.8077466], [60.4956612, 56.805435], [60.4997295, 56.8033912], [60.4987237, 56.8024818], [60.5088486, 56.8002759]], [[60.5384512, 56.7041597], [60.5396291, 56.7072872], [60.5157048, 56.7106853], [60.5088158, 56.7101643], [60.505726, 56.7038697], [60.5139558, 56.7046298], [60.5105042, 56.7001092], [60.5016302, 56.697431], [60.504651, 56.6950921], [60.5029734, 56.6924498], [60.5056071, 56.692065], [60.501341, 56.6898213], [60.506556, 56.688525], [60.5018831, 56.6850303], [60.5031971, 56.6827187], [60.5023441, 56.6808203], [60.5053686, 56.6804076], [60.509588, 56.6759242], [60.5181938, 56.6725782], [60.5431359, 56.6922208], [60.5390244, 56.6946366], [60.5410797, 56.6952674], [60.53999, 56.6958311], [60.5677539, 56.7084449], [60.5646311, 56.71109], [60.5616618, 56.7096931], [60.5510232, 56.7114904], [60.5529175, 56.7097994], [60.5431924, 56.7037366], [60.5413215, 56.7026486], [60.5384512, 56.7041597]], [[60.4712395, 56.7948051], [60.4715588, 56.796403], [60.4671115, 56.7960663], [60.4639148, 56.7939713], [60.4627771, 56.7893561], [60.4653165, 56.7884702], [60.4641135, 56.7874758], [60.4705793, 56.7865916], [60.470531, 56.7844611], [60.4780037, 56.7853363], [60.4838723, 56.791109], [60.4787117, 56.7937752], [60.4712395, 56.7948051]]]}, "bbox": [60.0070815, 56.5937449, 60.9432814, 56.9826914]}], "bbox": [60.0070815, 56.5937449, 60.9432814, 56.9826914]}'

m = folium.Map(location=[56.838924, 60.605701], zoom_start=9, tiles='OpenStreetMap')

geo_ekb = folium.GeoJson(data=polygon_ekb,
                              tooltip= 'Свердловская область<br>город Екатеринбург', #подсказки всплывающие при наведении мыши
                              style_function=lambda x: {'fillColor': 'blue', 'fillOpacity': 0.3, 'color': 'gray', 'weight': 3, 'opacity': 0.8})
folium.Popup('Численность населения — 1 495 066 чел.<br>').add_to(geo_ekb) # подсказки статика по клику мыши

geo_ekb.add_to(m)

#отображаем карту
m.save('results/map.html')
webbrowser.open('/results/map.html')
